pyproject.toml
[tool.poetry]
name = "timecontroller"
version = "0.1.0"
description = "Проект контроля времени доставки через удобные и простые интерфейсы"
authors = ["Garik <khvatik.igor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.0"
aiogram = "^3.13.1"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
